[tests] E: Refactoring Linux App
