+++ src/benchmarks/linux-app/src/identity.rs
@@ -1,0 +1,28 @@
+// Copyright(c) The Maintainers of Nanvix.
+// Licensed under the MIT License.
+
+//==================================================================================================
+// Imports
+//==================================================================================================
+
+use ::posix::unistd;
+
+//==================================================================================================
+// Standalone Functions
+//==================================================================================================
+
+pub fn test_getpid() {
+    // Try to get PID
+    match unistd::getpid() {
+        Ok(pid) => {
+            ::nvx::log!("got PID {:#?}", pid);
+        },
+        Err(err) => {
+            panic!("failed to get PID: {:?}", err);
+        },
+    };
+}
+
+pub fn test() {
+    test_getpid();
+}
